requirements.txt
streamlit
azure-identity
azure-ai-agents
azure-ai-inference
azure-ai-projects